+++ Services/Product/IProductService.cs
@@ -5,8 +5,8 @@ namespace Services.Product;
 public interface IProductService
 {
     IEnumerable<ProductDTO> GetProducts();
-    ProductDTO GetProduct(int id);
+    ProductDTO GetProduct(long id);
     void AddProduct(ProductDTO customer);
     void UpdateProduct(ProductDTO customer);
-    void DeleteProduct(int id);
+    void DeleteProduct(long id);
 }
